Added error messages to protocol (#39)
The Client / Server communication protocol now contains an error field that can be used to communicate that an error has occurred.
As an example of this, if the client does not encrypt a message for all of a recipients devices, the server will return an error stating the id's of the missing devices.